pyproject.toml
[tool.poetry]
name = "corvus-python"
version = "1.0.0"
description = "An interim home for data-related & python-based utility functionality"
authors = ["endjin <endjineers@endjin.com>"]
license = "Apache-2.0"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
azure-identity = "^1.15.0"
docxtpl = "^0.16.7"
pyodbc = "^5.1.0"
pyspark = { version = "^3.3.1", optional = true }
delta-spark = { version = ">=2.2.0,<4", optional = true }
azure-communication-email = "^1.0.0"
polars = "^1.38.1"
pandera = {extras = ["polars"], version = "^0.29.0"}
deltalake = "^1.4.2"
opentelemetry-api = "^1.40.0"
azure-storage-blob = "^12.28.0"
fastexcel = "^0.19.0"
fsspec = "^2026.2.0"
azure-data-tables = "^12.7.0"

[tool.poetry.extras]
pyspark = ["pyspark", "delta-spark"]

[tool.poetry.group.dev.dependencies]
jupyter = "~1.0.0"
pandas = ">=1.5.1,<3"
isort = "^5.13.2"
black = "^24.4.2"
mypy = "^1.11.1"
dummy-notebookutils = "^1.5.1"
pyspark = "^3.3.1"
delta-spark = ">=2.2.0,<4"

[tool.poetry.group.test.dependencies]
pytest = "^7.1.2"
pytest-cov = "^3.0.0"
behave = "v1.2.7.dev6"
junitparser = "^3.1.1"
dummy-notebookutils = "^1.5.1"
flake8 = "~6.1.0"
flake8-pyproject = "^1.2.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120

[tool.flake8]
exclude = [".git", "__pycache__", "dist", ".venv"]
max-line-length = 120
extend-ignore = ["E203", "E701"]

[tool.behave]
paths = ["./tests/bdd"]

[tool.isort]
profile = "black"
skip = ["__init__.py", ".git", "__pycache__", "dist", ".venv"]

[tool.mypy]
exclude = ["__init__\\.py$", '.git', '__pycache__', 'dist', '.venv']
ignore_missing_imports = true
disallow_untyped_defs = true
implicit_optional = true
